+++ objectbox/CHANGELOG.md
@@ -15,6 +15,8 @@
   MongoDB to match the name used in the MongoDB database. [#713](https://github.com/objectbox/objectbox-dart/issues/713)
 * Provide a helpful error message if the name of a property conflicts with a target ID property 
   created for a `ToOne` relation. [#713](https://github.com/objectbox/objectbox-dart/issues/713)
+* Generator: find `@Backlink()` source relation also in case target entity class is processed first.
+  [#687](https://github.com/objectbox/objectbox-dart/issues/687)
 
 ### Sync